+++ mail/mail-uniffi/CHANGELOG.md
@@ -11,6 +11,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Fix
 
   - `markMessageAsRead` now mark message as Read
+  - `LabelAs` now update only LabelType::Label
 
 ## [0.21.4] - 2024-11-13